New UI for Metamist (#291)
* Deploy to analysis.meta changes to main (#48)

* Improvements to analysis (#45)

* Add 'meta', 'active', 'on_behalf_of' to analysis schema

* Add 'exome' to sample_sequencing.type

* Add 'exome' to SequenceType

* Add 'active' and 'meta' to analysis models

* Implement 'meta', 'active', 'on_behalf_of' to analysis

* Add 'UNKNOWN' to analysis status

* Finish implementation of new analysis fields

* Fix default for AnalysisUpdateModel

* Remove _meta from incorrectly renamed method

* Fix "get_latest_analysis_for_samples_and_type" method

* Final fixes from tests

* Validate Backup (#28)

* Set up skeleton for test script

* Add unittests

* parameterize testing interface

* Add GCP Secret + Refactor
* Pulls GCP secret that stores prod + local connection details.
* Combines get_results and get_databases into one function
*Now raises runtime error if mysql dir is not empty.
* Creates connection in setUp function.

* Add restore script
* Restores DB to latest version uploaded in GCS (or specified folder)

* Add DR Documentation

* Fix indent

* Adjust indent

* Format Instructions

* Fix local links

* Add external linkss

* Update in line with code review + refactor

* Add requirements.txt

* Add link to image

* Update db/backup/README.md


* Set defaults for local configuration
* Pulls from environment variables rather than secret. Sets defaults

* Update local config set up.

* Add further detail to the TODO

* Add TOC

* re-write recovery procedure

* Update validation instructions

* Delete yearly validation plan
* Without the deletion of the VM image, it's the same as the biannual plan

* Add SM Validation

* Apply suggestions from code review


* Modify service account


* Globally generic metadata parser (#46)

* Initial commit for generic metadata parser

* Add QC analysis to generic parser

* Allow input key maps + add qc parsing

* Improve sequence matching in parser

* Improve metadata collapsing logic


* Fix license description (#182)

* First (mega) commit

* Missed a dependency

* Implement search

* Fix persistent error message + remove participant_in_seqr

* Fill in extra statistics

* Linting

* Remove package-lock from git

* Restore package-lock to old version

* Minor improvements to display + linting

* Fix test

* Added Sample and Participant models from API

* Added URL redirects to history, change external to internal sample ID in URL

* Added whitespace margins to sample page

* Added sequence interface from API, error handling search, narrow width info page

* Removed old css file

* Added icons to search

* switched to semantic-ui and removed reactstrap

* Added searchResultToRoute function with alerts for unimplemented cases

* Added new Links file with Sample, Family and Participant links

* Change page size to a query param

* Fix linting

* Updated npm, changes search result CSS + removed alerts for unimplemented pages

* Fix minor analysis stats

* Minor fix to search route handling + image size

* Release (#301)

* Update test subset file (#295)

* Fixes to create_test_subset
* Allow for participant transfer without ped
* Also modifies how random families are selected
* Switch seed back
* fix comment
* Add extra comments
* Add comments + fixes to get_random_families
* Fix _get_random_families logic
* Add type hints

* Fix Get Sequences End Point  (#298)

* Bump version: 5.0.0 → 5.0.1 (#300)

* Add muck the duck

* Fix searching non-existent sample by ID

* Escape query term

* Fixed page reset bug when changing projects

* Allow server to decide heading names for ProjectSummary

* Fix linting

* Fix test

* Added catch for external IDs on sample page

* Add tests

* Removed console.log()

* Add more tests (and fix them too)

+++ test/testbase.py
@@ -10,6 +10,7 @@ from functools import wraps
 
 from typing import Dict
 
+from pymysql import IntegrityError
 from testcontainers.mysql import MySqlContainer
 import nest_asyncio
 
@@ -163,9 +164,10 @@ class DbTest(unittest.TestCase):
 
     @classmethod
     def tearDownClass(cls) -> None:
-        db = cls.dbs[cls.__name__]
-        db.exec(f'DROP DATABASE {db.MYSQL_DATABASE};')
-        db.stop()
+        db = cls.dbs.get(cls.__name__)
+        if db:
+            db.exec(f'DROP DATABASE {db.MYSQL_DATABASE};')
+            db.stop()
 
     def setUp(self) -> None:
         self.project_id = 1
@@ -185,4 +187,8 @@ class DbIsolatedTest(DbTest):
         for table in TABLES_ORDERED_BY_FK_DEPS:
             if table in ignore:
                 continue
-            await self.connection.connection.execute(f'DELETE FROM {table}')
+            try:
+                await self.connection.connection.execute(f'DELETE FROM {table}')
+                await self.connection.connection.execute(f'DELETE HISTORY FROM {table}')
+            except IntegrityError as e:
+                raise IntegrityError(f'Could not delete {table}') from e
